+++ main.go
@@ -77,6 +77,7 @@ func main() {
 		publicAuth.GET("/stats/user", middleware.RequirePermission("read:athlete-unknown:user-stats"), server.GetUserStats)
 		publicAuth.POST("/stats/user/migrate", middleware.RequirePermission("migrate:athlete-unknown:user-stats"), server.MigrateUserStats)
 		publicAuth.GET("/upcoming-rounds", middleware.RequirePermission("read:athlete-unknown:upcoming-rounds"), server.GetUpcomingRounds)
+		publicAuth.PUT("/user/username", middleware.RequirePermission("update:athlete-unknown:profile"), server.UpdateUsername)
 	}
 
 	// Admin endpoints (API key auth)